+++ neovim/lua/mini_config.lua
@@ -86,25 +86,8 @@ function M.setup()
 	end, { desc = "Toggle mouse", unique = true })
 
 	require("mini.completion").setup({
-		delay = { completion = 1000, info = 1000, signature = 1000 },
+		delay = { completion = 150, info = 150, signature = 150 },
 	})
-
-	-- local align_blame = function(au_data)
-	-- 	if au_data.data.git_subcommand ~= "blame" then
-	-- 		return
-	-- 	end
-	--
-	-- 	-- Align blame output with source
-	-- 	local win_src = au_data.data.win_source
-	-- 	vim.wo.wrap = false
-	-- 	vim.fn.winrestview({ topline = vim.fn.line("w0", win_src) })
-	-- 	vim.api.nvim_win_set_cursor(0, { vim.fn.line(".", win_src), 0 })
-	--
-	-- 	-- Bind both windows so that they scroll together
-	-- 	vim.wo[win_src].scrollbind, vim.wo.scrollbind = true, true
-	-- end
-	-- local au_opts = { pattern = "MiniGitCommandSplit", callback = align_blame }
-	-- vim.api.nvim_create_autocmd("User", au_opts)
 end
 
 return M